test(cleanup): refactor deploy token fixture
